requirements.txt
altair==5.0.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click @ file:///C:/ci_311/click_1676433091657/work
colorama @ file:///C:/ci_311/colorama_1676422310965/work
contourpy==1.0.7
cycler==0.11.0
Flask @ file:///C:/ci_311/flask_1676436667658/work
Flask-SQLAlchemy==3.0.3
fonttools==4.39.4
greenlet==2.0.2
idna==3.4
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///C:/ci_311/jinja2_1676424968965/work
jsonschema==4.17.3
kiwisolver==1.4.4
MarkupSafe @ file:///C:/ci_311/markupsafe_1676424152318/work
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psycopg2==2.9.6
pyparsing==3.0.9
pyrsistent==0.19.3
pyTelegramBotAPI==4.11.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML @ file:///C:/ci_311/pyyaml_1676432488822/work
requests==2.30.0
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.14
telegram==0.0.1
toolz==0.12.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug @ file:///C:/b/abs_8578rs2ra_/croot/werkzeug_1679489759009/work
